pyproject.toml
[project]
name = "bang-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "bhyun08",email = "nbhyun0329@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "uvicorn (>=0.34.0,<0.35.0)",
    "pymysql (>=1.1.1,<2.0.0)",
    "fastapi (>=0.115.6,<0.116.0)",
    "sqlalchemy (>=2.0.37,<3.0.0)",
    "cryptography (>=44.0.0,<45.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
